// DigitalCafedra/Lab7/enums.ts
export enum Docs {
    passport,
    driverCard,
    snils,
    birthCertificate
}

export enum CarsBody {
    SEDAN = 'Sedan',
	HATCHBACK = 'Hatchback',
	COUPE = 'Coupe',
	MINIVAN = 'Minivan',
}

export enum CarsClass {
    A = 'A',
	B = 'B',
	C = 'C',
	D = 'D',
	E = 'E',
	F = 'F',
}